refactor: 타입 구조 개선 및 level을 difficulty로 변경
- Category 관련 타입을 별도 파일로 분리
- card.level을 card.difficulty로 변경
- Meaning과 Difficulty의 유니온 타입으로 Category 생성
- isLevel 헬퍼 함수를 getCategoryType으로 변경
- FSRS 관련 프로퍼티를 FSRSParameters 타입으로 그룹화
